Combined TabPanel.jsx into Tabs.jxs
