+++ bundles/admin/admin-layereditor/components/Tabs.jsx
@@ -3,6 +3,8 @@ import PropTypes from 'prop-types';
 import AntTabs from 'antd/lib/tabs';
 import 'antd/lib/tabs/style/css';
 
+const AntTabPane = AntTabs.TabPane;
+
 export const Tabs = ({children, ...other}) => {
     return (
         <AntTabs {...other}>
@@ -11,6 +13,18 @@ export const Tabs = ({children, ...other}) => {
     );
 };
 
+export const TabPane = ({children, ...other}) => {
+    return (
+        <AntTabPane {...other}>
+            {children}
+        </AntTabPane>
+    );
+};
+
 Tabs.propTypes = {
+    children: PropTypes.any
+};
+
+TabPane.propTypes = {
     children: PropTypes.any
 };
